+++ src/Easy.h
@@ -101,6 +101,7 @@ class Easy : public Nan::ObjectWrap {
   static NAN_METHOD(Send);
   static NAN_METHOD(Recv);
   static NAN_METHOD(Perform);
+  static NAN_METHOD(Upkeep);
   static NAN_METHOD(Pause);
   static NAN_METHOD(Reset);
   static NAN_METHOD(DupHandle);