+++ CHANGELOG.md
@@ -70,12 +70,14 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
   - `CURLOPT_SSH_COMPRESSION`
   - `CURLOPT_TLS13_CIPHERS`
   - `CURLOPT_TIMEVALUE_LARGE`
+  - `CURLOPT_UPKEEP_INTERVAL_MS`
 - Add missing info fields: 
   - `CURLINFO_*_{DOWNLOAD,UPLOAD}_T`
   - `CURLINFO_*_TIME_T`
   - `CURLINFO_FILETIME_T`
 - Add `Curl.getVersionInfo()` which returns an object that represents the struct returned from `curl_version_info()`
   See their type definition for details: [`./lib/types/CurlVersionInfoNativeBinding.ts`](./lib/types/CurlVersionInfoNativeBinding.ts)
+- Add `upkeep` function to Easy and Curl classes. This is a binding for the `curl_easy_upkeep()` function.
 - Errors thrown inside callbacks are correctly caught / passed forward (if using multi interface)
 - Added **experimental** `curl.<http-verb>()` async api  
   This API can change between minor releases.